SESSIONS IS GOOD ENOUGH
At this point I have sessions fully working with the correct datatypes for corresponding areas.
Is it done? No, there is still more work that could be done, but at this time it will stay as is for the foreseeable future.
This update fixes sessions, so now you can use the view all page to see all your sessions, add a new session with corrected information and datatypes, and return to the view all sessions page to see your newly inserted session.